requirements.txt
streamlit==1.33.0
requests
python-dotenv